update the version

+++ .install/displaymanager.sh
@@ -70,33 +70,49 @@ if [ "$dmsel" == "Install sddm and theme" ] ;then
 
     disman=0
     # Try to force the installation of sddm
-    echo ":: Install sddm"
-    yay -S --noconfirm sddm sddm-sugar-candy-git --ask 4
-
+    echo ":: Installing sddm"
+    sudo pacman -S --noconfirm --needed sddm qt5-graphicaleffects qt5-quickcontrols2 qt5-svg --ask 4
+    
     # Enable sddm
     if [ -f /etc/systemd/system/display-manager.service ]; then
         sudo rm /etc/systemd/system/display-manager.service
     fi
     sudo systemctl enable sddm.service
 
-    if [ ! -d /etc/sddm.conf.d/ ]; then
-        sudo mkdir /etc/sddm.conf.d
-        echo "Folder /etc/sddm.conf.d created."
-    fi
+    echo 
+    if gum confirm "Do you want to install the sddm-sugar-candy theme?" ;then
+        echo ":: Installing sddm-sugar-candy-git"
 
-    sudo cp sddm/sddm.conf /etc/sddm.conf.d/
-    echo "File /etc/sddm.conf.d/sddm.conf updated."
+        if [ -d ~/Downloads/sddm-sugar-candy ] ;then
+            rm -rf ~/Downloads/sddm-sugar-candy
+            echo ":: ~/Downloads/sddm-sugar-candy removed"
+        fi 
+        wget -P ~/Downloads/sddm-sugar-candy https://github.com/Kangie/sddm-sugar-candy/archive/refs/heads/master.zip
+        echo ":: Download of sddm-sugar-candy complete"
+        unzip -o -q ~/Downloads/sddm-sugar-candy/master.zip -d ~/Downloads/sddm-sugar-candy
+        echo ":: Unzip of sddm-sugar-candy complete"
+        sudo cp -r ~/Downloads/sddm-sugar-candy/sddm-sugar-candy-master /usr/share/sddm/themes/sugar-candy
+        echo ":: sddm-sugar-candy copied to target location"
 
-    if [ -f /usr/share/sddm/themes/sugar-candy/theme.conf ]; then
+        if [ ! -d /etc/sddm.conf.d/ ]; then
+            sudo mkdir /etc/sddm.conf.d
+            echo "Folder /etc/sddm.conf.d created."
+        fi
 
-        # Cache file for holding the current wallpaper
-        sudo cp wallpapers/default.jpg /usr/share/sddm/themes/sugar-candy/Backgrounds/current_wallpaper.jpg
-        echo "Default wallpaper copied into /usr/share/sddm/themes/sugar-candy/Backgrounds/"
+        sudo cp sddm/sddm.conf /etc/sddm.conf.d/
+        echo "File /etc/sddm.conf.d/sddm.conf updated."
 
-        sudo cp sddm/theme.conf /usr/share/sddm/themes/sugar-candy/
-        sudo sed -i 's/CURRENTWALLPAPER/'"current_wallpaper.jpg"'/' /usr/share/sddm/themes/sugar-candy/theme.conf
-        echo "File theme.conf updated in /usr/share/sddm/themes/sugar-candy/"
+        if [ -f /usr/share/sddm/themes/sugar-candy/theme.conf ]; then
 
+            # Cache file for holding the current wallpaper
+            sudo cp wallpapers/default.jpg /usr/share/sddm/themes/sugar-candy/Backgrounds/current_wallpaper.jpg
+            echo "Default wallpaper copied into /usr/share/sddm/themes/sugar-candy/Backgrounds/"
+
+            sudo cp sddm/theme.conf /usr/share/sddm/themes/sugar-candy/
+            sudo sed -i 's/CURRENTWALLPAPER/'"current_wallpaper.jpg"'/' /usr/share/sddm/themes/sugar-candy/theme.conf
+            echo "File theme.conf updated in /usr/share/sddm/themes/sugar-candy/"
+
+        fi
     fi
 
 elif [ "$dmsel" == "Deactivate current display manager" ] ;then
